+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 - Code blocks now use the syntax highlighting theme's caret and selection colors.
 - `CommonMarkCache` now implements `Debug` ([#7](https://github.com/lampsitter/egui_commonmark/pull/7) by [@ChristopherPerry6060](https://github.com/ChristopherPerry6060)).
-- `CommonMarkCache::add_syntax_theme_from_folder`
+- `CommonMarkCache::add_syntax_themes_from_folder`
+- `CommonMarkCache::add_syntax_theme_from_bytes`
 
 ### Changed